@@ -15,14 +15,14 @@
   "private": true,
   "dependencies": {
     "@angular-devkit/schematics": "^17.3.7",
-    "@angular/animations": "^17.3.9",
-    "@angular/common": "^17.3.9",
-    "@angular/compiler": "^17.3.9",
-    "@angular/core": "^17.3.9",
-    "@angular/forms": "^17.3.9",
-    "@angular/platform-browser": "^17.3.9",
-    "@angular/platform-browser-dynamic": "^17.3.9",
-    "@angular/router": "^17.3.9",
+    "@angular/animations": "^18.0.0",
+    "@angular/common": "^18.0.0",
+    "@angular/compiler": "^18.0.0",
+    "@angular/core": "^18.0.0",
+    "@angular/forms": "^18.0.0",
+    "@angular/platform-browser": "^18.0.0",
+    "@angular/platform-browser-dynamic": "^18.0.0",
+    "@angular/router": "^18.0.0",
     "@canvasjs/angular-charts": "^1.0.3",
     "bootstrap": "^5.3.1",
     "bootstrap-icons": "^1.9.1",
@@ -43,8 +43,8 @@
     "@angular-eslint/schematics": "17.5.1",
     "@angular-eslint/template-parser": "17.5.1",
     "@angular/cli": "^17.3.7",
-    "@angular/compiler-cli": "^17.3.9",
-    "@angular/language-service": "^17.3.9",
+    "@angular/compiler-cli": "^18.0.0",
+    "@angular/language-service": "^18.0.0",
     "@fortawesome/fontawesome-free": "^5.15.4",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
@@ -66,7 +66,7 @@
     "run-script-os": "^1.1.6",
     "ts-node": "^8.10.2",
     "tslint": "^6.1.3",
-    "typescript": "~5.2.2",
+    "typescript": "~5.4.5",
     "webpack-subresource-integrity": ">=1.5.1",
     "websocket-extensions": ">=0.1.4"
   }
